implementation/discord-bot-python/requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
requests==2.32.5
better-profanity==0.7.0
